converted use time hook, to context provider and adjusted passing of state to components
